Update memory manager module colors and icons
Now, the colors of the icons for the memory and I/O blocks changes
with the theme. In addition, the text and the color of the memory
module icons in the memory manager can be selected with CSS.

In order to be able to change the color of the icons, instead of
using external SVG files, which cannot access CSS variables, they
had to be defined as strings in icons.js.